Document operators and punctuation
Summary:
Document all Thrift operators and punctuation in a dedicated section of the IDL reference replacing the Separators section. The old wording didn't make clear which lexical elements can be separated with comma and semicolon and incorrectly stated that we use commas in examples (we follow the VSCode formatting rules).

Replace "punctuators" with "punctuation" in the lexer for consistency.

This completes the cleanup of the Lexical Elements section.

#forcetdhashing

Reviewed By: avalonalex

Differential Revision: D43848725

fbshipit-source-id: d921d5b925febdac7b9077ec428160f9b12e4235