Disallow pathological qualified identifiers
Summary:
Disallow periods not followed by a valid identifier start character in qualified identifiers, e.g. `foo.` (notice the trailing period) is no longer permitted. This brings the lexer closer to the actual grammar of qualified identifiers documented in the IDL reference as

```
identifier ("." identifier)*
```

Technically this is a breaking change but since such pathological qualified identifiers don't work in most contexts already this is unlikely to cause real breakage.

#forcetdhashing

Reviewed By: iahs

Differential Revision: D43879630

fbshipit-source-id: 02f119f0b0e6d8fa6d197ee371931a09d3318de2