Use third-party/tbb to build HHVM (#9344)
Summary:
We currently don't use the TBB in third-party/tbb for building HHVM,
which is causing some issues with the Ubuntu 22.04 update.

This PR updates our CMake files to use the TBB in third-party. This leverages TBB's
`tbb_build` macro that builds TBB as part of the CMake configuration step.

This uses a custom rule for building TBB instead of using the built-in `findTBB.cmake`
logic as I couldn't get it to play nicely with a custom tbb path. We may want to revisit
this once we rolled onto newer OSes though as `find_package` is the recommended
way to depend on the library, even when using `tbb_build`.

Pull Request resolved: https://github.com/facebook/hhvm/pull/9344

Reviewed By: hgoldstein

Differential Revision: D43825878

Pulled By: Wilfred

fbshipit-source-id: b9c823f7a52d6505d0949c49f653d0960fea7722