Fix bug related to uneven argument and parameter lists
Summary:
We were zipping parameter and argument lists before checking for subtypes, but these lists need not have the same length because there might be default parameters. Consequently, we were emitting no constraints with the assumption that we are in an impossible state.

This diff fixes the problem by first curtailing the parameter list before zipping.

Reviewed By: mheiber

Differential Revision: D43908153

fbshipit-source-id: e4698eeef4d5a17e1486723a892cc084e61d86a7