Coalesce should conditionally refine in right argument
Summary:
For reasons unknown, we were typechecking coalesce as if right side is guaranteed to be evaluated.

This diff fixes it by saving the environment after evaluating the left argument and unioning it with the state after evaluating the right.

Implementation was also explicitly generating fresh type variables where there isn't an explicit need. I just let environment and type union logic handle it in the background.

Differential Revision: D43920551

fbshipit-source-id: 975cd88ec29a206750fff07389ddc45b71d388bb