Flip `const_default_func_args` and `const_default_lambda_args` to on by default
Summary:
In preparation for removing these flags and inlining the behavior, flip the default to true and fix up / remove tests.

Generally, I left parser and typechecker tests with the errors intact, and just deleted runtime tests that weren't doing anything beyond testing patterns that would fail to parse in the future.

Reviewed By: Wilfred

Differential Revision: D43888421

fbshipit-source-id: 556bf9f6ec01c4b31b2e6ad0bb3f7595f9505faa