Generate ast and nast modules in oxidized
Summary:
The ast and nast modules are intended to be convenience modules providing type aliases for the AST/NAST instantiations of the generic AAST types (both instantiate the annotated AST with unit as the expression-annotation and unit as the environment-annotation). We currently have the Nast module hand-written in OCaml (synced via hh_oxidize) and the `ast` module hand-written in Rust. This allows for them to get out of sync with the AST (if some type is added to `ast_defs` or `aast_defs`, but not nast.ml or ast.rs).

Instead, generate the ast and nast modules from the type definitions in `ast_defs` and `aast_defs`.

Reviewed By: shayne-fletcher

Differential Revision: D43924433

fbshipit-source-id: c32aafcd747bbbc74164a06849ef454635e0fa9b