Use nast module
Summary:
The `Transform` and `Pass` traits make heavy use of inlining for efficiency of fused passes. This makes compilation of uses of the traits expensive, so we do not expose them. Since we do not expose them, and we only ever use them on named ASTs, restrict them so that they only operate on NASTs. This reduces annotation burden in elab passes by quite a bit, and hopefully makes the logic in this crate easier to follow.

We may revert this change if we end up needing to use the `Transform` and `Pass` traits for transformations of TASTs (or some new instantiation of the AAST) in the future, but since we don't have any plans for this at the moment, I think it's reasonable to simplify for now.

Reviewed By: shayne-fletcher

Differential Revision: D43924448

fbshipit-source-id: 6338260fc3d803c6dc922900a107ff9f405ceac4