Lwt_utils.exec_checked supports cancellation
Summary:
`Lwt_utils.exec_checked` is the helper we use in our codebase to launch a process under Lwt, and capture its output. It was introduced in D13541290.

For HackIDE, when it shells out to "hh --find-refs" or "hh check", we'll want (1) asynchronous so that HackIDE can continue to do work while the shellout happens, (2) cancellable so that the shell-out doesn't do any work.

When D13541290 introduced `exec_checked`, it came with the comment that upon Lwt.cancel cancellation, "this function will not kill the underlying process. (I tried to implement it, but after killing the process, both Lwt_io.close and Lwt_io.abort would hang when trying to close the process's stdin/stdout/stderr."

Well, I still want cancellation, so that's what this diff provides.

The ocaml guidance on how to do cancellation:
* https://github.com/ocsigen/lwt/issues/283/#issuecomment-518014539 "All variants of cancel I know of, including the Lwt one, complicate programming. Users rarely understand what cancel really does... Cancel doesn't do what its name seems to suggest... Cancel breaks the nice property that an initial promise can only be resolved through its resolver... Based on the above, I think we should not try to fix cancel, but discourage using it, and leave it alone. For now, I think it's better to ask people to do cancelation explicitly."
* https://github.com/ocsigen/lwt/blob/master/src/core/lwt.mli `val cancel : _ t -> unit` It is recommended to avoid [Lwt.cancel], and handle cancelation by tracking the needed extra state explicitly within your library or application.

So I used the C# approach, which is to explicitly pass a "cancellation-token" into the function and use it to request cancellation. I used `~(cancel_token : unit Lwt.t)` for the cancellation-token, and listen for cancellation requests like this:
```
let%lwt () = Lwt.choose [
  let%lwt () = cancel_token in process#kill Sys.sigkill; Lwt.return_unit);
  let%lwt _ = process#status in Lwt.return_unit); ]
```
In other words: wait until cancellation is requested and thereupon send sigkill, but stop waiting once the process completes. This avenue didn't encounter any of the problems that D13541290 had encountered. I added some explanation:
```
There are several ways for it to be cancelled:
* [cancel] parameter
  e.g. `let (cancel, u) = Lwt.bind(); Lwt.wakeup_later u (); exec_checked ~cancel`
  Upon the cancel parameter being completed, we will send a SIGKILL to the underlying
  process. In Unix, this cannot be ignored nor blocked. It will result in
  Process_failure.status=WSIGNALED Sys.sigkill, and .stdout/.stderr will contain
  whatever the process had sent before that.
* [timeout] parameter
  e.g. `exec_checked ~timeout:30.0`
  This parameter is passed straight through to Lwt_process.open_process_full.
  Its behavior is to aggressively closes stdout/stderr pipes and also send a SIGKILL.
  It will result in Process_failure.status=WSIGNALED Sys.sigkill, and .stdout/.stderr
  will be empty because of the closed pipes, and .exn will likely be something about
  those closed pipes too.
* Lwt.cancel
  e.g. `Lwt.pick [exec_checked ...; Lwt.return_none]`.
  This has been deprecated as a cancellation mechanism by the Lwt folks.
  This will *NOT* kill the underlying process. It will close the pipes.
  As to what happens to the promise? you shouldn't really even be asking,
  and indeed varies depending on the process. *)
```

Note: for sake of tests, I also added `Exec_command.Shell`.

Reviewed By: nt591

Differential Revision: D43850687

fbshipit-source-id: 8f798d0929e0666e163941b7f421d765b005bbe6