fix eager import of elements when performing hierarchy checks in traits
Summary:
HHVM ignores `require class` constraints when flattening traits.  As such, on the `methods_trait_conflict_01.bad.php` example below it fatals when loading `TParentDemo` because it sees conflicting definitions of `asEnt` and it is not aware that the conflict will be resolved when `TParentDemo` is flatted inside the `Demo` class.

Hack instead, when performing hierarchy checks of traits with `require class` constraints performs an eager element fetch from the required class.  This is important to avoid false errors when the trait implements multiple interfaces defining the same method with different signatures, and the actual method implementation is provided by the require class (see test `hack/test/requireclass/typing//typing/methods_philippe_01.good.php`).

This diff synchronises Hack with the HHVM behaviour: the eager element resolution is enabled only to resolve ambiguities if the parent is an interface, not if it is a trait.

Reviewed By: vsiles

Differential Revision: D43943837

fbshipit-source-id: 8bc5347e99932f0a937bdeaf6fdaf4ed50c6f25a