Call into elaborate_namespaces_visitor in hackc elab
Summary:
In the hackc elaborate command, use the elaborate_namespaces_visitor to elaborate type names in method and function bodies.

In the long run, we may pull the elaborate_namespaces_visitor pass into the elab crate (just as the Naming module performs Naming_elaborate_namespaces_endo, then the lambda capture pass, then the elab passes, in a single API). But they're separate for now (and kind of need to be, since we haven't ported the lambda capture pass yet).

Reviewed By: shayne-fletcher

Differential Revision: D43922879

fbshipit-source-id: 055e1425fe70a8c4235518a5d6b94adce27c154a