Intersect inside of Awaitable with enum base type
Summary: Under implicit pessimisation, we pessimise an enum of the form `enum E : int as int` as `(~E & int)`. But we get this wrong with `Awaitable<E>`, intersecting the `Awaitable` itself and so producing `nothing`. Fix code so that we pessimise the inside of the `Awaitable` in the case that the element type is enforced (for async functions).

Reviewed By: vassilmladenov

Differential Revision: D44021186

fbshipit-source-id: fd91e442f26a57a0cdb3dcb0a6ef1c1ccff93c9e