load LocalConfig from /etc/hh.conf
Summary:
Properly initializing a GlobalOptions object requires reading values from both
`<root>/.hhconfig` and `/etc/hh.conf`. After a lot of concentrating and reverse
engineering, it seemed like the cleanest thing to do was for HhConfig to parse both.

This diff adds a handwritten LocalConfig type that is a subset of ServerLocalConfig.t.
It's handwritten, because using hh_oxidize on ServerLocalConfig.t is blocked by the
use of module types, and because only a few GlobalConfig options are even relevant.

Later diffs will add more options on need-driven basis.

The ".hhconfig" path was hardcoded in several places; use the hh_config as a
single source of truth where possible.

Reviewed By: nt591

Differential Revision: D44013589

fbshipit-source-id: 7bef1bc080a090736e43081db0cd7d6402178180