progress 3/4 lifetime of progress.json
Summary:
Before this diffstack, ServerProgress was only ever specific to the PID of a server instance and was only read while a connection attempt was pending. Therefore, the only invariant we needed was "the ServerProgress must give correct information for any time a connection is pending".

This diffstack has changed ServerProgress to be about the repo as a whole, so we need to extend the invariant - ServerProgress must now be correct for ANY time. I think the follow areas cover it.
1. If a monitor has never been started - it says "unknown hh_server state"
2. As soon as the monitor starts - now says "monitor initializing..."
3. If the monitor terminates cleanly through unhandled exception or Exit.exit - ServerProgress file gets deleted, and it says "unknown hh_server state"
4. If the monitor terminates abruptly e.g. due to SIGKILL - I changed the ServerProgress file to also mention the PID who wrote it, and if the PID is dead then `ServerProgress.read` will now return "unknown hh_server state".
5. If a server is dead but monitor is still alive - "server stopped", or some other more specific server-exit-status message if the server happened to leave one in its finale file
6. If a server is alive and has a connection pending - same invariant as before
7. If a server is alive and typechecking with no connection pending - this was already handled before
7. If server has no connection pending and is not typechecking - it now shows "ready". (it used to show "HackIDE:active" due to an oversight).

There's a small race window: if the most recent progress message came from server, and server dies, then it will briefly have "unknown hh_server_state" before monitor reports "server stopped".

Reviewed By: nt591

Differential Revision: D43991705

fbshipit-source-id: ec17c8263a4aada813c495cf21e734806be76d5c