progress 1/4 progress.json now based on root, not pid
Summary:
It used to be that we'd write progress to `/tmp/hh_server/progress.<PID>.json`, so a given server instance would only write to its own PID-specific progress file. How did "hh_client check" learn that PID? ... it had to wait until it had finished a handshake with the monitor, since the monitor is the first place it could learn the server's PID from. This also meant we had no way for hh_client to display monitor progress e.g. when it was stuck.

This diff changes it to `/tmp/hh_server/zWWW/progress.json`, so the server and monitor can both write to a repo-specific progress file. This relies on the fact (enforced by `/tmp/hh_server/zWWW.lock`) that only a single monitor and hence single server instance ever exist for a repo.

The benefit from this design:
* hh_client will be able to display progress while it's waiting for the monitor
* clientLsp will be able to display progress while waiting for a shell-out, without needing to interact with the monitor to obtain PID
* (and in future I intend to use this design for `/tmp/hh_server/zWWW/errors`, incremental errors, so clients can `tail -f` those errors also without needing a connection)
* (however, I have left `/tmp/hh_server/<PID>.fin`, the finale file, as being PID-specific. This is the file where a server process will write its last gasp information, detailed information about the callstack and error code which led it to terminate. The finale file is read by hh_client to learn why the connection it had to the server just died. I think it makes sense for this to be PID-specific)

My diffstack plan:
1. This diff changes the progress file to be based on root
2. The next diff eliminates "warning" progress, dead code
3. The third diff deals with the lifetime semantics of the progress file, i.e. when it's created and when it expires
4. The fourth diff cleans up the messages

**WHAT THIS DIFF DOES**
1. The progress file is now root-related. It was too hard to thread root through every call to ServerProgress (there are many of them all through the typechecker, in places that don't have root easily accessible). Instead I make ServerProgress.root be a global mutable variable, which I expect to be initialized at the start of a process' lifetime; subsequent attempts to read or write progress will throw an exception if it hasn't been initialized. The only process which read or write progress are (1) hh_client, (2) monitor, (3) server, (4) some integration tests that exercise the serverMain loop directly.
2. It used to be that writing to the progress file was split into ServerProgress.ml and ServerCommandTypes.ml. I moved it all into ServerProgress.ml

Reviewed By: nt591

Differential Revision: D43968881

fbshipit-source-id: 0f0aa48e785eecbd8eaec0686370af21ed6198b2