progress 2/4 no more warning
Summary:
This diff makes only one change. In the ServerInit.ml snippet below, it removes the call to `ServerProgress.send_warning`. This is the only call to `send_warning` in the entire codebase.
```
let lazy_saved_state_init () =
  match ServerLazyInit.saved_state_init ... with
  | Ok _ -> ...
  | Error err ->
    let message = load_state_error_to_verbose_string err in
    let (next_step_descr, next_step) =
      if not genv.local_config.SLC.require_saved_state then
        ("fallback", Exit_status.No_error)
      else if auto_retry then
        ("retry", Exit_status.Failed_to_load_should_retry)
      else
        ( "fatal", Exit_status.Failed_to_load_should_abort)
    in
    (match next_step with
    | Exit_status.No_error ->
      ServerProgress.send_warning user_message
```
The warning was for the scenario where hh_server attempts a native saved-state-load, but it fails, and hh.conf "require_saved_state=false", so it falls back to a full init. The user experience in this case was to display a warning message:
```
$ ... tweak .hhconfig
$ hh --config require_saved_state=false
hh_server is busy: [waiting on db...] /
Could not load saved-state from DevX infrastructure. The underlying error message was: Hack can't start because .hhconfig isn't recognized.
If you are editing .hhconfig, sorry, Hack does not support starting up while there is an edited .hhconfig.
Otherwise, are you in the middle of a rebase? Can you `arc pull`?
Once done, restart Hack by clicking on it in the status bar or running hh restart on the command line.
If that fails, post in Hack Language Feedback/Support.

hh_server is busy: [indexing - this can take a long time, see warning above]
```

In other words, the text that is usually an error would now be displayed as a warning, in addition to the usual progress message `hh_server is busy: [indexing]`.

But our hh.conf.erb makes require_saved_state=true in all cases, so we never go down this path. Therefore, I deleted last remaining call to `ServerProgress.send_warning`, and I deleted this API entirely. There were a whole load of knock-on cleanups that I could therefore do throughout the rest of the codebase.
* ServerProgress used to support calling "send_warning" at one time, and "send_progress" at another, so it had to maintain a state of what the most recent warning and most recent progress messages where, so they could be combined together when it writes the progress.json file
* clientConnect used to support reading the warning and progress, and appending "this can take a long time", and checking whether the warning has changed
* If you look at ServerProgress.mli, the API has been simplified. I also renamed "ServerProgress.send_progress" to just ServerProgress.write, because that's what it does.

Reviewed By: nt591

Differential Revision: D43969343

fbshipit-source-id: 7d41880a57f214a94fa737ca42c444650bc449c8