EL.log_subsampling
Summary:
I wanted some way for HackEventLogger to be throttled - to let us write events safe in the knowledge that even if there's some kind of storm then it won't overburden the telemetry infrastructure.

We already have "has_logged_invariant_violation_bug" which throttles with the rule "no more than once per typecheck". But I wanted a more general purpose form of throttling, one that would be good in a long-running process like the monitor. This is in preparation for the next diff in the stack, which indeed adds telemetry to the monitor.

```
(** This function is to subsample events, in case there might be too many of them
and you don't want to flood scuba. But for well-behaved events it will still record every one.
Even if it has to subsample, it still records how many samples were dropped in a "subsample_drop_count" field.
  EL.get_sample "EVENT" |> C.data "a" |> EventLogger.log_subsampling
or,
  EL.get_sample "EVENT" |> C.data "a" |>
  EventLogger.log_subsampling ~subsample_key:"EVENT_a" ~max_tolerable_long_term_interval:60.0 ~max_tolerable_short_term_spike:30
Over the long term, it won't write to scuba more than 1 event per 60.0s (default).
Over the short term, it wo't write to scuba more than 30 events in a spike (default).
It does this analysis on a per-key basis (which defaults to the event name).

e.g. if we had a chronic event that fired once a second, then it'd allow
through the first 30 when the arrived, but thereafter it'd only record one a minute;
each one's "subsample_drop_count" woud be 60, indicating that for each one recorded
there were a further 60 that weren't.

e.g. if we had a busy loop with 1k events/s, the first 30 would be recorded. Subsequently it would
record only 1 every 60s.

e.g. if we had an event like typechecking that happened say five times in a minute because the user
was making changes, but afterwards it was idle for five minutes, then the "backlog" of events
would be cleared. *)
```

Reviewed By: madgen

Differential Revision: D44016342

fbshipit-source-id: b7d287a519d7651b2929c5a5d479fba2631d4f44