Noop refactor: Reduce nonlocal mutation in arg parsing
Summary:
We had a pattern where we would use action-at-a-distance to work with `Args.Tuple`.
I think it's easier to read and maintain to keep the refs local where possible.

Also eta-reduce `fun x -> foo := x` to `( := ) ref` in arg parsing: I think this makes it clearer when we're doing logic vs. no-logic.

Reviewed By: vassilmladenov

Differential Revision: D44100584

fbshipit-source-id: 29ddcc2b50b9dbda695fc630c6f13f931c632a41