add BucketHashSelector for bucketized PoolRoutes
Summary:
special hash selector that does the selection based on a bucket id in the fiber context.
IMPORTANT: the bucket id in the fiber context must be present otherwise the selection throws. This means, the bucketized PoolRoute must be wrapped in McBucketRoute that computes the bucket id and adds it into the fiber context.

Next diff: adding bucketId to a request in `DestinationRoute`

Reviewed By: stuclar

Differential Revision: D44109030

fbshipit-source-id: 7a3a494859cddef54559b53491aa6540fbd635ca