Keep `{ and }` on the same line for Expression Tree Blocks
Summary:
## Context
For Expression Tree Blocks hackfmt would format the code so that the backtick and open/closing braces would be on separate lines like:
```
Code`
  {
    $bar = 4;
    return $bar;
  }
`;
```

Based on user feedback the preferred formatting would keep the backtick and braces on the same line like:

```
Code`{
  $bar = 4;
  return $bar;
}`;
```

## Solution
For `PrefixedCodeExpression` I do an additional match on `body` to determine if it is a `Syntax.CompoundStatement`. If so I invoke `handle_compound_statement`.

To get the formatting I wanted, I added an option to `handle_compound_statement` to avoid prepending a `Space` before the compound statement. Otherwise the formatting would turn into:

```
Code` {
  $bar = 4;
  return $bar;
}`;
```

Reviewed By: madgen

Differential Revision: D44176824

fbshipit-source-id: 0ab6e008aa836338400894aa9b30b2f07179541c