compute set of files referenced through xrefs
Summary:
Add an option  to the indexer so that it produces the set of files directly referenced through xrefs, from the set of input files.

Main changes:
1. add an option (as a config parameter)
2. add a path field in `Symbol_sym_def` (the datatype that represents symbol definitions).
3. Union the referenced files produced by each worker (by augmenting the JobReturn.t type

We only need to index the www files, the hhi file are embedded in the indexer and we always can produce facts for them with the `symbol_write_include_hhi` option.

Motivation [diff navigation/glass snapshot](https://fb.quip.com/Vu8GAJ1n69xY)

Reviewed By: simonmar

Differential Revision: D42992958

fbshipit-source-id: a3ff1e9d019fb8445558397d914ac0f175f3a85d