angery_reaccs_only
Summary:
Here's a tiny bit of refactoring. It used to be that the function `ClientConnect.tty_progress_reporter` made an internal decision (based on the presence of /tmp/hh_server/ANGERY_REACCS_ONLY) for whether or not to use emojis for the progress-spinner.

While streaming errors are being rolled out, I wanted a fun whimsical facebook-spirited indicator that errors are indeed streaming. That way the early adopter volunteers will recognize that they're seeing something fun. I figured I could re-use our angery reaccs. (I'll switch it back to normal once the feature is stable and ready for wider rollout).

This diff therefore allows the caller to decide whether to use angery reaccs for the spinner.

Reviewed By: hgoldstein

Differential Revision: D44159581

fbshipit-source-id: 8030810d71b86deb65579d30cad321ac24c78701