expose Marshal_tools.preamble
Summary:
The "preamble" is used for RPC between client and monitor/server. The pattern there is to send the size of a chunk in a five-byte word with parity byte, followed by a marshalled blob of that size.

I want to use the same approach for the errors.bin streaming error file. So I'm exposing `Marshal_tools.preamble`

Reviewed By: CatherineGasnier

Differential Revision: D44159583

fbshipit-source-id: 2d4f43d6edd15cb9522c0fa26dd27e6707d62c39