Proc.is_alive
Summary:
For the "per-repo" status files (progress.json, errors.bin) we need to deal with the fact that the creating process might be killed uncleanly, e.g. with SIGKILL, leaving the file behind abandoned. How can a consumer determine that the file is no longer valid?

The pattern I've adopted is (1) the creating process writes its PID into the file, (2) the consuming process will read the PID and determine whether that PID is still alive. This diff adds a library function `Proc.is_alive pid` to do this check.

```
(** There is no reliable way to tell whether an arbitrary PID is still alive.
That is because after the process has long since died, its PID might be recycled
for use by another process. This routine makes a best-effort attempt:
1. Can we read /proc/<pid>/cmdline? - yes if it is alive or still being held onto.
2. Is "expected" a substring of cmdline? - this is our best-effort protection
against pid recyling, since the chance that it gets recycled *and* the recycled one
has the process-name we were expecting just seems pretty low. Note that cmdline
will be empty if the process has been zombified. Also that if a process modifies
its argv then the cmdline will reflect that. In effect, cmdline is the command-line
that the process wants you to see.
3. Can we send SIGKILL 0 to it? If this succeeds without throwing an exception,
then the process is alive. *)
```

{F908227867}

Reviewed By: CatherineGasnier

Differential Revision: D44159586

fbshipit-source-id: 1339f3ff2bb6c039d46e9efa5f51ab6b2cd5afa2