ServerProgress.mli doc-comments
Summary:
Just adding some comments to ServerProgress.mli. I wanted to document the lifecycle-semantics of the progress.json file.

```

(** Progress is a file in /tmp/hh_server/<repo>.progress.json which is written
by monitor+server. It lives from the moment the monitor starts up until the
moment it finally dies or is killed. You should only read it by the `read`
call, since that protects against races. Anyone at any time can read this
file to learn the current state. The state is represented solely as a
human-readable string to be shown to the user in the CLI or VSCode status bar.
It specifically should not be acted upon in code -- it is slightly handwavey
in places (e.g. there is an interval from when a server dies until the monitor
realizes that fact where attempting to read will say "unknown"). *)
```

Reviewed By: CatherineGasnier

Differential Revision: D44159587

fbshipit-source-id: cc5a3b7acd0e42d0a3e7bfe04dfea71a4ff00b10