Add BlobEncoder support for HHBBC type system
Summary:
Add serialization/deserialization support for the HHBBC type system
(and associated classes). For the most part, this is
straight-forward. There's one gotcha: A serialized res::Class is
always unresolved, even if it was resolved when serialized. This is
because resolution of a res::Class requires an Index, and one might
not available when deserializing. There are various schemes one can
use to resolve while deserializing, but they're brittle and
complicated.

Instead we'll always get an unresolved class and it must be manually
resolved afterwards (if necessary). To accomplish this, a new function
called resolve_classes() is provided.

Reviewed By: jano

Differential Revision: D44236493

fbshipit-source-id: a6f552c8ff1fd1e030a474ba843f417ab949ed7d