Pre-initialize function return types using extern-worker
Summary:
As part of the initial "flatten" pass, also calculate the initial
return type for every function. This is easy to do because
type-constraints are now guaranteed to be fully resolved at this
point.

Add a FuncInfo2 class, which stores the equivalent of FuncInfo but in
extern-worker jobs. FuncInfo2s for methods are stored on their
declaring class, and FuncInfo2s for top-level functions are stored on
their own. As with ClassInfo2, we convert FuncInfos2 into FuncInfos
inside make_local() and proceed as normal. This takes advantage of the
newly added Type serialization. We resolve any classes once the
FuncInfo vector is built.

Reviewed By: jano

Differential Revision: D44237947

fbshipit-source-id: dcc324d6dc6ef11173e89c2717969d5188c777ee