Invalidate folded descendant classes
Summary: We correctly invalidate the shallow and folded decls of classes in files which changed, but we neglected to add Rust_provider_backend behavior for this site where we invalidate descendant folded classes of the classes which changed. This resolves the Decl_heap_elems_bug issue which caused us to roll back last week.

Reviewed By: shayne-fletcher

Differential Revision: D44277395

fbshipit-source-id: 6e9ccddd6a33b561211fb9dc02f138b9480e8504