Add syntax for soft includes for packages
Summary:
This diff just adds soft_includes into the package.toml syntax and a basic test related to it. It does not add any error checks for deployments yet; for now, this series of diffs will just allow us to separate the error codes in the typechecker, but we'll add runtime support for this in the coming weeks.

Note that the runtime doesn't need to know anything about soft_includes: only soft deploys.

Differential Revision: D44271438

fbshipit-source-id: 91636881bb44c4a5ee0fcbf16ff123c1d79b27df