Add typing errors for soft include, make separate error codes for package errors
Summary:
This diff splits the typing errors for package stuff into their own error codes. Soft includes throws the same error as hard includes, just with slightly different wording and a different error code.

The reason to split the error codes in the typechecker is for migration purposes. If we used the same code for both, then if someone FIXMEs or ignores one error for some reason, they won't get any feedback if we change a package from being soft included to not included at all, even though that has actual runtime implications.

Differential Revision: D44271440

fbshipit-source-id: 4a11a59c11e9cc54e06ef9a936fdd30128920105