Array assignment ordering
Summary:
In expressions of the form `e1[e] = e2`, we were typechecking `e1` before `e`. HHVM does it the other way around. This makes the implementation both unsound and incomplete.

The fix is simply to change the order of typechecking.

See the added test cases for repros of unsoundness and incompleteness that is now fixed.

Reviewed By: Wilfred

Differential Revision: D44261184

fbshipit-source-id: ee6933dbb14c50dca03cca0d4efcba221057261f