Don't mutate programs before codemod
Summary:
This function is only called by codemod programs, which definitely don't want to run mutators on the way to modifying the original file.
This reveals a bug in the name implementation for placeholder typedef, which is fixed.
Also remove the need to pass the useless --gen flag to the codemod programs.

Reviewed By: Mizuchi

Differential Revision: D44277269

fbshipit-source-id: 796137fa3efb6da94e3078ed072fa6fd406c4a60