Factor out binary ops to allow explicit transforms of lhs / rhs expressions
Summary:
This a fairly churny refactor the AST which pulls out the payload of the `Binop` contructor into to its own record type. Why do this?

If you look at many of our existing elaboration & validation passes over binary expressions, we often want to set the context differently for the lhs expression, particularly when the operation in question is assignment.

When working with visitors we can do this straightforwardly by modifying context only when recursing into that expression.

With the `ppx_transform` approach (and equivalent Rust version), we have sacrificed this fine grained control of recursion in exchange for the ability to fuse passes. To acheive the same thing in the transform approach we need to be able to annotate the lhs expression so that we can transform explicitly i.e. to make additional changes to the context over those made for an expression.

The changes in this diff allow us to do just that.

Reviewed By: rjbailey

Differential Revision: D44260664

fbshipit-source-id: c0be80cbc2612be106d86f0c0dd664c18ce6b13d