Remember origins of shape types
Summary:
Shape type can get huge; however in source code
they are often referred to via aliases. This diff
remembers the aliases to speed up comparison of
shape types. This new information could also be
used to display better error messages.

One could think that D32675495 got most of the
wins to be had here, but that is not the case. I think
that is because comparing types that come from
distinct localize calls does not fall in the physical
equality fast path.

Reviewed By: vsiles

Differential Revision: D43848923

fbshipit-source-id: 55cf4d924288133ca8835674287b03c96d0f0c4c