Always create a default async resource pool
Summary:
I’ve discussed this in separate thread with you guys, as we decided that for now we believe that it should. Reasons:

> One of the reasons we might need a default is so we can route our internal methods (e.g. FacebookBase2). Alternatively, we could route those to a separate resource pool.

> ThriftServer has lifecycle hooks like onStartServing() and onStopRequested(). It also has an AsyncScope that can run background tasks. We run all of this on an executor that is ontained from getHandlerExecutorKeepAlive(). So I think we would always need an executor that ThriftServer can use for these purposes. We can’t randomly pick a RP’s executor. So I’m inclining towards always enforcing defaultAsync() to be always provided.

FYI Sri is working on making this “right” by introducing selectResourcePool customization point that selects RP for a given request.

We've decided  that pragmatic solution for now is to always set up default async RP.

Reviewed By: praihan

Differential Revision: D44205495

fbshipit-source-id: 0bec5fbc3ab45ce771568e6f46af59f37fc73b22