Check invariants using extern-worker
Summary:
Move the check_invariants step in HHBBC to run within extern-worker
jobs. We cannot verify everything we used to because it relied on
global knowledge (plus some of the data structures are different).

Keep the old check_invariants pass, but rename it
check_local_invariants and only run it after make_local. Once
make_local is removed, it will be removed as well.

Reviewed By: kavoor

Differential Revision: D44325845

fbshipit-source-id: 3c299417e39e890e0139b6a467a31a501bb97ee2