Add lightweight testing for refactoring code actions
Summary:
## What

Add a lightweight test for the method override refactoring code action which was introduced in D32997165.

A refactoring code action (our terminology) is a code action that does not fix an error.

## How

The test works with new infra for hh_single_typecheck-driven tests for refactoring code actions.

Our test files must contain a range specified with special markers: `/*range-start*/` and `/*range-end*/`: this enables us to avoid hard-coding locations, making tests easy to generate and evolve.

## Why now

I'm working on "extract variable" and similar refactors and want to be able to write cheap and easy-to-read tests to enable fast iteration and easy reviews.

## Advice requested

Suggestions welcome if you can think of a way to make the .exp file easier to read.

Reviewed By: ljw1004

Differential Revision: D44293523

fbshipit-source-id: 33d2385c99c02dd049e10cd0bd90e229bda6e996